config: adding the concept of a boot_firmware setting
boot_firmware contains information for loading a default bitstream on startup: paths, handles and flags
If bitstream stuff is specified it will be loaded first.
If overlay stuff is specified, it will be applied next.
If not enough information is provided, it will skip each step.
Both steps will be attempted.

Signed-off-by: Artie Poole <stuart.poole@canonical.com>