config: user config completely overrides vendor config
previously, the configs would be merged but this merging was just concatenating vectors
this meant that any duplicate loads would be called multiple times.
In an attempt to find a smart way to merge these, it became clear that actually the user
probably only wants one source of truth for default bitstreams.
User can specify empty arrays to disable defaults like
```
[boot_firmware]
bitstreams = []
overlays = []
```

Signed-off-by: Artie Poole <stuart.poole@canonical.com>