platforms: apply and load now check for and strip longest fw_source_path
the source paths are retrieved as lines in
`/sys/module/firmware_class/parameters/path` with `/lib/firmware`
appended to the end.

These are sorted to be longest first, and the first match to the source
path is stripped from the front to get the shortest possible relative path.

This relative path is written to the relevant control file.

Signed-off-by: Artie Poole <stuart.poole@canonical.com>