platforms: change how matching substrings works for platforms
xilinx platform, for example, will be capable of controlling versal, zynq and zynqmp so
instead of matching lots of full platform strings to the same type, the map is reversed
to match all (split on commas) provided substrs to a platform.

e.g. if a user provides "xlnx,versal-fpga", this will only ever match to Xilinx platform

Signed-off-by: Artie Poole <stuart.poole@canonical.com>