root: app: update snapcraft file to expose cli and dbus daemon (#52)
Updated snapcraft file as explained in title. `cli` application frontend
is available and we dont need backend to be completed to start exposing
it. In addition to that I also updated the daemon configuration so
system dbus service is up and running when the fpgad snap is installed.
I used `busctl` to check if it accessible and it looks like the
connection is there and busctl is able to find the application in
classic ubuntu installations.

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>