platforms: softeners: use PLATFORM_REGISTRY instead of substrings (#65)
* platforms: re-organize functions

just a re-organization of the functions in the module no functionality
change.

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>

* platforms: softeners: use PLATFORM_REGISTRY instead of substrings

This replaces PLATFORM_SUBSTRINGS logic with platform registry. Which
relies on one to one match for given comma separated list of compatible
strings. It also provides easy to extend interface for new platforms to
be added.

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>

* platforms: softeners: remove `platform_type` api

looks like we dont really need `platform_type` to be accessed, it is not
being used anywhere so I will remove it. If we see a need for it we can
create it again.

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>

* app: softeners: add missing license headers

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>

* platforms: add simple test for match_platform_string

Added a basic test to make sure that compat string matching is working
as expected in good weather scenario.

Also added `Any` trait super to make sure the platforms can be
downcasted.

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>

---------

Signed-off-by: Talha Can Havadar <talha.can.havadar@canonical.com>