cli: add support for loading files from current directory (#121)
* cli: add support for loading files from current directory

Signed-off-by: Artie Poole <stuart.poole@canonical.com>

* dbus: comm: remove the path checking on load calls

the path should not be accesible to a strictly confined snap,
it is a miracle it was working before, but it appears as though
this hole has been patched and now this fails with
`Call failed: FpgadError::Argument: <path> is not a valid path to a bitstream file.`
unless it is stored in a path accessible via the connected plugs
i.e. kernel-firmware-control providing /lib/firmware access.

Signed-off-by: Artie Poole <stuart.poole@canonical.com>

* root: add sandbox dir to .gitignore

helps with keeping, e.g. local version of provisioning
yaml files

Signed-off-by: Artie Poole <stuart.poole@canonical.com>

* tests: fix integration tests now that paths cannot be inspected

since the "path.exists" returns false due to confinement now
(snapd changes?) the tests have been edited to account for
these changes

Signed-off-by: Artie Poole <stuart.poole@canonical.com>

---------

Signed-off-by: Artie Poole <stuart.poole@canonical.com>